refactor: extract student_profile panel from PanelViews.tsx, closing #144
Final batch of the per-panel split for issue #144: student_profile -
539 lines, the largest panel in the codebase, with 7 local useState
pairs and the only call site of usePanelData's updateStudentLocally
mutator (built in PR 1 for exactly this purpose). PanelViews.tsx is
now just the router shell.

- New panels/StudentProfilePanel.tsx ({ students, schools, reportsList,
  currentUser, token, updateStudentLocally } props). Moves in all 7
  local state pairs (sel, profileTab, editingProfile, profileDraft,
  savingProfile, searchQuery, showDropdown, activityFilter) - confirmed
  via grep every one is used exclusively by this panel - plus the
  sel-initializing useEffect that defaults to the first student.
- PanelViews.tsx's 1 if-block replaced with a one-line router call
  passing all 6 props.
- Final unused-import cleanup: with this the last panel moved,
  PanelViews.tsx no longer needs apiFetch, useState/useEffect, Table/
  Column, MetricCard, PageHeader/EmptyStudents, handleDownloadPDF, any
  of its 16 lucide icons, or 6 of 8 imported types (Student, ClassGroup,
  School, Worksheet, LogEntry, Ticket all confirmed dead - only User
  and UserRole remain in use). PanelViews.tsx is down to 92 lines: the
  usePanelData call, ~20 one-line router dispatches, and the
  still-untouched dead CONTENT_ITEMS array flagged back in the original
  #144 design proposal.

Hit and fixed a genuine TypeScript inference quirk during extraction:
Object.entries(r.conceptMastery).forEach(([topic, mastery]) => {...})
compiled clean on unmodified main (confirmed by re-running tsc against
a fresh pre-PR8 checkout) and in an isolated 10-line repro of the exact
same construct, but failed with "Type 'unknown' is not assignable to
type 'string'" inside the real 571-line extracted file. Root cause
appears to be TypeScript's inference giving up under this file's depth/
complexity (it's the single largest component in the codebase) rather
than any actual logic difference - the byte-for-byte code is identical
to what shipped on main for weeks. Fixed with a minimal, purely
type-level annotation, ([topic, mastery]: [string, string]), zero
runtime behavior change. Verified live afterward that the exact
downstream code path (Skill Proficiency section, Recommended Focus
Areas) renders correctly with real data.

Verified: tsc --noEmit clean, vite build succeeds, byte-diff of the
moved block (plus the one intentional type annotation). Thorough live
click-through against a scratch backend with a real seeded evaluation
report: searchable student-selector dropdown (sel/searchQuery/
showDropdown), switching between students, all 4 tabs (Overview -
Score Trend/Skill Proficiency/Recommended Focus rendering correctly
with the type-fixed code path, Academic Record - real report data,
Personal Details - edited Blood Group through the UI, confirmed
updateStudentLocally fired and the PATCH persisted to MongoDB,
Activity Log - activityFilter toggle narrows the list correctly).

This is the last panel - #144 (PanelViews.tsx split into per-panel
files) is now fully resolved.
